checkpoint

+++ implot_demo.cpp
@@ -460,7 +460,7 @@ void ShowDemoWindow(bool* p_open) {
             if (ImPlot::IsPlotHovered() && ImGui::IsMouseClicked(0) && ImGui::GetIO().KeyCtrl) {
                 ImPlotPoint pt = ImPlot::GetPlotMousePos();
 
-                data.push_back(ImVec2(pt.x, pt.y));
+                data.push_back(ImVec2((float)pt.x, (float)pt.y));
             }
             ImPlot::PushStyleVar(ImPlotStyleVar_Marker, ImPlotMarker_Diamond);
             if (data.size() > 0)
@@ -719,8 +719,8 @@ void ShowDemoWindow(bool* p_open) {
     if (ImGui::CollapsingHeader("Offset Data")) {
         float xs[50], ys[50];
         for (int i = 0; i < 50; ++i) {
-            xs[i] = 0.5 + 0.4 * cos(i/50.f * 6.28);
-            ys[i] = 0.5 + 0.4 * sin(i/50.f * 6.28);
+            xs[i] = 0.5f + 0.4f * cos(i/50.f * 6.28f);
+            ys[i] = 0.5f + 0.4f * sin(i/50.f * 6.28f);
         }
         static int offset = 0;
         ImGui::SliderInt("Offset", &offset, -100, 100);
